refactor(content): extract ./constants into /libs/env
