+++ content/index.js
@@ -4,10 +4,10 @@ const {
   CONTRIBUTOR_SPOTLIGHT_ROOT,
   REPOSITORY_URLS,
   ROOTS,
-  VALID_LOCALES,
   HTML_FILENAME,
   MARKDOWN_FILENAME,
-} = require("./constants");
+} = require("../libs/env");
+const { VALID_LOCALES } = require("../libs/constants");
 const Document = require("./document");
 const Translation = require("./translation");
 const { getPopularities } = require("./popularities");
